actually renamed de_solvers to de in this commit, changed necessary references so that tests pass again, and updated test_type_utils.py to import from the right place
