+++ qiskit/providers/aer/pulse/controllers/mc_controller.py
@@ -22,9 +22,9 @@ import time
 import numpy as np
 from scipy.linalg.blas import get_blas_funcs
 from qiskit.tools.parallel import parallel_map, CPU_COUNT
-from ..de_solvers.pulse_de_solver import construct_pulse_zvode_solver
-from ..de_solvers.pulse_utils import (cy_expect_psi_csr, occ_probabilities,
-                                      write_shots_memory, spmv_csr)
+from ..de.pulse_de_solver import construct_pulse_zvode_solver
+from ..de.pulse_utils import (cy_expect_psi_csr, occ_probabilities,
+                              write_shots_memory, spmv_csr)
 
 dznrm2 = get_blas_funcs("znrm2", dtype=np.float64)
